app/requirements.txt
Django==4.2.2
gunicorn==20.1.0
djangorestframework==3.14.0
selenium==4.10.0
beautifulsoup4==4.12.2
lxml==4.9.2
